pyproject.toml
[project]
name = "pharma-doc-validator-mcp"
version = "0.1.0"
description = "MCP server for pharmaceutical document validation (SOP logic, ELN, acceptance criteria)"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-docx>=1.1.0",
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
pharma-validator = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
